6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ int main(void)
 {
 	printf("Size of char: %lu bytes\n", sizeof(char));
 	printf("Size of int: %lu bytes\n", sizeof(int));
-	printf("Size of long: %lu bytes\n", sizeof(long int));
+	printf("Size of long: %lu bytes\n", sizeof(int));
 	printf("Size of long long: %lu bytes\n", sizeof(long long int));
 	printf("Size of float: %lu bytes\n", sizeof(float));
